Move the `testing-client` module under the `@bayou` namespace.
