+++ local-modules/@bayou/testing-client/EventReporter.js
