+++ local-modules/testing-server/EventReceiver.js
@@ -7,7 +7,7 @@ import { CommonBase } from '@bayou/util-common';
 import TestCollector from './TestCollector';
 
 /**
- * Receiver of events sent by {@link testing-client.EventReporter}.
+ * Receiver of events sent by {@link @bayou/testing-client.EventReporter}.
  */
 export default class EventReceiver extends CommonBase {
   /**
